AB#94093 - preparation for staging, unification in the solution
