+++ src/HE.CRM.AHP.Plugins/HE.CRM.AHP.Plugins.csproj
@@ -119,7 +119,6 @@
     <Compile Include="Handlers\CustomApi\LocalAuthority\GetMultipleLocalAuthoritiesHandler.cs" />
     <Compile Include="Handlers\CustomApi\SendReminderEmailForFinaliseDraftApplicationHandler.cs" />
     <Compile Include="Handlers\CustomApi\SendReminderEmailForRefferedBackToApplicantHandler.cs" />
-    <Compile Include="Handlers\CustomApi\SendRequestToAssignContactToExistingOrganisationHandler.cs" />
     <Compile Include="Handlers\CustomApi\SentInviteContactToJoinOrganisationHandler.cs" />
     <Compile Include="Handlers\CustomApi\SetApplicationHandler.cs" />
     <Compile Include="Handlers\CustomApi\SetDeliveryPhaseHandler.cs" />
@@ -217,7 +216,6 @@
     <Compile Include="Plugins\CustomApi\GetSingleHomeTypePlugin.cs" />
     <Compile Include="Plugins\CustomApi\GetTypeOfHomesListPlugin.cs" />
     <Compile Include="Plugins\CustomApi\SendReminderEmailForRefferedBackToApplicantPlugin.cs" />
-    <Compile Include="Plugins\CustomApi\SendRequestToAssignContactToExistingOrganisationPlugin.cs" />
     <Compile Include="Plugins\CustomApi\SentInviteContactToJoinOrganisationPlugin.cs" />
     <Compile Include="Plugins\CustomApi\SetApplicationPlugin.cs" />
     <Compile Include="Plugins\CustomApi\SetDeliveryPhasePlugin.cs" />
