+++ src/HE.CRM.AHP.Plugins/Services/Contacts/IContactService.cs
@@ -5,7 +5,6 @@ namespace HE.CRM.AHP.Plugins.Services.Contacts
 {
     public interface IContactService : ICrmService
     {
-        void SendRequestToAssignContactToExistingOrganisation(Guid organisationId, Guid contactId);
         void InviteContactToJoinExistingOrganisation(Guid inviterId, Guid invitedContactId, Guid organisationId);
     }
 }
