+++ src/HE.CRM.Plugins/Services/Contacts/IContactService.cs
@@ -1,3 +1,4 @@
+using System;
 using HE.Base.Services;
 using HE.Common.IntegrationModel.PortalIntegrationModel;
 
@@ -9,5 +10,6 @@ namespace HE.CRM.Plugins.Services.Contacts
         ContactRolesDto GetContactRoles(string email, string ssid, string portal);
         ContactDto GetUserProfile(string contactExternalId);
         void UpdateUserProfile(string contactExternalId, string serializedContact);
+        void SendRequestToAssignContactToExistingOrganisation(Guid organisationId, Guid contactId);
     }
 }
