+++ src/HE.CRM.Plugins/Services/GovNotifyEmail/GovNotifyEmailService.cs
@@ -287,5 +287,43 @@ namespace HE.CRM.Plugins.Services.GovNotifyEmail
                 _ = _loanApplicationRepositoryAdmin.ExecuteGovNotifyNotificationRequest(govNotReq);
             }
         }
+
+
+        public void SendNotifications_COMMON_REQUEST_TO_ASSIGN_CONTACT_TO_EXISTING_ORGANISATION(EntityReference contactId, EntityReference associatingContactId)
+        {
+            if (contactId != null && associatingContactId != null)
+            {
+                var contact = _contactRepositoryAdmin.GetById(contactId.Id, nameof(Contact.OwnerId).ToLower(), nameof(Contact.FullName).ToLower(), nameof(Contact.EMailAddress1).ToLower());
+                var associatingContact = _contactRepositoryAdmin.GetById(associatingContactId.Id, nameof(Contact.FullName).ToLower());
+
+                if (contact != null && contact.OwnerId != null && associatingContact != null)
+                {
+                    this.TracingService.Trace("COMMON_REQUEST_TO_ASSIGN_CONTACT_TO_EXISTING_ORGANISATION");
+                    var emailTemplate = _notificationSettingRepositoryAdmin.GetTemplateViaTypeName("COMMON_REQUEST_TO_ASSIGN_CONTACT_TO_EXISTING_ORGANISATION");
+                    var govNotParams = new COMMON_REQUEST_TO_ASSIGN_CONTACT_TO_EXISTING_ORGANISATION()
+                    {
+                        templateId = emailTemplate?.invln_templateid,
+                        personalisation = new parameters_COMMON_REQUEST_TO_ASSIGN_CONTACT_TO_EXISTING_ORGANISATION()
+                        {
+                            recipientEmail = contact.EMailAddress1,
+                            username = contact.FullName ?? "NO NAME",
+                            subject = emailTemplate.invln_subject,
+                            associatinguser = associatingContact.FullName ?? "NO NAME",
+                        }
+                    };
+
+                    var options = new JsonSerializerOptions
+                    {
+                        Encoder = JavaScriptEncoder.UnsafeRelaxedJsonEscaping,
+                        WriteIndented = true
+                    };
+
+                    var parameters = JsonSerializer.Serialize(govNotParams, options);
+                    this.SendGovNotifyEmail(contact.OwnerId, contact.ToEntityReference(), emailTemplate.invln_subject, parameters, emailTemplate);
+                }
+            }
+        }
+
+
     }
 }
