+++ src/HE.CRM.Plugins/Services/GovNotifyEmail/IGovNotifyEmailService.cs
@@ -13,5 +13,6 @@ namespace HE.CRM.Plugins.Services.GovNotifyEmail
         void SendNotifications_INTERNAL_LOAN_APP_OWNER_CHANGE(invln_Loanapplication loanApplication, string subject, string appId);
         void SendNotifications_EXTERNAL_KYC_STATUS_CHANGE(Contact contact, string subject, Account organisation);
         void SendNotifications_EXTERNAL_APPLICATION_CASHFLOW_REQUESTED(invln_Loanapplication loanApplication);
+        void SendNotifications_COMMON_REQUEST_TO_ASSIGN_CONTACT_TO_EXISTING_ORGANISATION(EntityReference contactId, EntityReference associatingContactId);
     }
 }
